==> notebooks/01_data_translation.ipynb <==
import pandas as pd
import sys
import os

es_data = pd.read_excel("../private_data/data/survey_results_es/es-data.xlsx")

# column rename
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "../scripts/aggregation/")))
from rename_columns_es import rename_columns_es

es_data = rename_columns_es(es_data)

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "../private_data/dictionaries_student_data")))
from dictionary_students_es import es_student_to_cohort, es_student_to_id

# Adding "cohort_start_month" columns 
def student_to_cohort_func(each_name):
  return es_student_to_cohort[each_name]

student_cohorts = es_data["name"].apply(student_to_cohort_func)
es_data.insert(loc=2, column="cohort_start_month", value=student_cohorts) 

# Adding "cohort_language" column with "ES" value for Spanish cohorts
es_data.insert(3, "cohort_language", "ES")

# Adding "ct_student_id" columns
# es_data.insert(2, "ct_student_id", None) 
def student_to_id_func(each_name):
  return es_student_to_id[each_name]

# es_data["ct_student_id"] = es_data["name"].apply(student_to_id_func)
student_ids = es_data["name"].apply(student_to_id_func)
es_data.insert(loc=2, column="ct_student_id", value=student_ids) 

# dropping private and unneeded columns
es_data.drop("Timestamp", axis=1, inplace=True) # remove timestamp
es_data.drop("name", axis=1, inplace=True)
es_data.drop("final_comments", axis=1, inplace=True)

# pre_bootcamp_coding_experience replace
pre_bootcamp_coding_experience_dict = {
  "No tenía ningún conocimiento en programación": "I had zero coding knowledge",
  "Tenía conocimientos básicos de programación": "I had basic coding knowledge",
  "Tenía conocimientos entre medios y avanzados de programación": "I had medium to advanced coding knowledge",
  "Tenía conocimientos y experiencia profesional en programación": "I had professional coding knowledge and experience"
}
es_data['pre_bootcamp_coding_experience'] = es_data['pre_bootcamp_coding_experience'].replace(pre_bootcamp_coding_experience_dict)

# post_bootcamp_situation replace
post_bootcamp_situation_dict = {
  "Encontré un trabajo a tiempo completo como desarrollador (o un puesto similar)": "I found a Full-Time job as a developer (or a similar role)",
  "Encontré un trabajo a tiempo parcial como desarrollador (o un puesto similar)": "I found a Part-Time job as a developer (or a similar role)",
  "Encontré trabajo como desarrollador freelance, con una fuente de ingresos relevante": "I found work as a Freelance developer, with a relevant source of income",
  "Encontré un trabajo, pero no está relacionado con el bootcamp": "I found a job, but is not related to the bootcamp",
  "Aun no he encontrado trabajo": "I haven't found a job."
}
es_data['post_bootcamp_situation'] = es_data['post_bootcamp_situation'].replace(post_bootcamp_situation_dict)

# job_country replace
es_data['job_country'] = es_data['job_country'].str.strip()

job_country_dict = {
  "España": "Spain",
  "Alemania": "Germany",
}
es_data['job_country'] = es_data['job_country'].replace(job_country_dict)

# time_to_job replace
time_to_job_dict = {
  "Menos de 2 meses": "Less than 2 months",
  "Entre 3 y 4 meses": "3-4 months",
  "Entre 5 y 6 meses": "5-6 months",
  "Más de 6 meses": "More than 6 months",
}
es_data['time_to_job'] = es_data['time_to_job'].replace(time_to_job_dict)

# did_bootcamp_helped replace
did_bootcamp_helped_dict = {
  "Si": "Yes",
}
es_data['did_bootcamp_helped'] = es_data['did_bootcamp_helped'].replace(did_bootcamp_helped_dict)

# job_success_factors replace
#TODO PENDING. 
# job_success_factors_dict = {
#   "Buena comunicación y habilidades blandas": "Good communication and soft skills",
# }
# es_data['job_success_factors'].replace(job_success_factors_dict)

# is_job_hunting replace
is_job_hunting_dict = {
  "Sí, sigo buscando activamente un trabajo como desarrollador": "Yes, I keep actively looking for a job as a developer",
  "Sí, pero para ser honesto, estoy poniendo poco esfuerzo": "Yes, but to be honest, I'm putting little effort",
  "No, ya no estoy buscando trabajo como desarrollador": "No, I'm not looking for a job as a developer anymore",
}
es_data['is_job_hunting'] = es_data['is_job_hunting'].replace(is_job_hunting_dict)

# is_job_hunting replace
is_job_hunting_dict = {
  "Sí, sigo buscando activamente un trabajo como desarrollador": "Yes, I keep actively looking for a job as a developer",
  "Sí, pero para ser honesto, estoy poniendo poco esfuerzo": "Yes, but to be honest, I'm putting little effort",
  "No, ya no estoy buscando trabajo como desarrollador": "No, I'm not looking for a job as a developer anymore",
}
es_data['is_job_hunting'] = es_data['is_job_hunting'].replace(is_job_hunting_dict)

# job_hunt_blockers replace
#TODO PENDING. 

# search_country replace
es_data['search_country'] = es_data['search_country'].str.strip()

search_country_dict = {
  "España": "Spain",
  "Espanya": "Spain",
  "Alemania": "Germany",
  "Irlanda": "Ireland",
  "Europa": "Europe",
}
es_data['search_country'] = es_data['search_country'].replace(search_country_dict)

# company_interview_qty replace
company_interview_qty_dict = {
  "No tuve entrevistas o sólo muy cortas": "I didn't have interviews or just very short ones",
  "1 empresa": "1 company",
  "2 empresas": "2 companies",
  "3 empresas": "3 companies",
  "4 empresas": "4 companies",
  "5 empresas": "5 companies",
  "6 empresas": "6 companies",
  "7 empresas": "7 companies",
  "8 empresas": "8 companies",
  "9 empresas": "9 companies",
  "10 o más empresas": "10 or more companies",

}
es_data['company_interview_qty'] = es_data['company_interview_qty'].replace(company_interview_qty_dict)

# job_search_week_hours replace
job_search_week_hours_dict = {
  "Menos de 5 horas Semanales": "Less than 5",
  "Entre 5 y 10 horas Semanales": "Between 5 and 10",
  "Entre 10 y 20 horas Semanales": "Between 10 and 20",
  "Entre 20 y 30 horas Semanales": "Between 20 and 30",
  "Más de 30 horas Semanales": "More than 30",
}
es_data['job_search_week_hours'] = es_data['job_search_week_hours'].replace(job_search_week_hours_dict)

# coding_week_hours replace
coding_week_hours_dict = {
  "Menos de 5 horas Semanales": "Less than 5",
  "Entre 5 y 10 horas Semanales": "Between 5 and 10",
  "Entre 10 y 20 horas Semanales": "Between 10 and 20",
  "Entre 20 y 30 horas Semanales": "Between 20 and 30",
  "Más de 30 horas Semanales": "More than 30",
}
es_data['coding_week_hours'] = es_data['coding_week_hours'].replace(coding_week_hours_dict)

# job_search_focus replace
job_search_focus_dict = {
  "Me concentré principalmente en mejorar en JavaScript simple y en la resolución de problemas (por ejemplo, katas)": "I focused mostly on getting better at plain JavaScript and problem solving (e.g. katas)",
  "Me concentré principalmente en mejorar en otras tecnologías que conozco (por ejemplo, React, Express)": "I focused mostly on getting better at other technologies that I know (e.g React, express)",
  "Me concentré principalmente en aprender nuevas tecnologías": "I focused mostly on learning new technologies",
  "Hice un poco de todo (JavaScript, frameworks, nuevas tecnologías)": "I did a bit of everything (JavaScript, frameworks, new technologies)",
  "Ninguna de las anteriores": "None of them",
}
es_data['job_search_focus'] = es_data['job_search_focus'].replace(job_search_focus_dict)

#NOTE: cambiar casing the technologies directo en excel. todo lowercase y separado con comas. Quitar extra.
# ie: java, php, python

# new_projects_qty replace
new_projects_qty_dict = {
  "No, no hice ningún nuevo proyecto": "No, I haven't",
  "Si, hice 1 nuevo proyecto": "Yes, I created 1",
  "Si, hice de 2 a 3 nuevos proyectos": "Yes, I created 2 to 3",
  "Si, hice más de 3 nuevos proyectos": "Yes, I created more than 3",
}
es_data['new_projects_qty'] = es_data['new_projects_qty'].replace(new_projects_qty_dict)

# did_bootcamp_helped replace
did_create_portfolio_dict = {
  "Si": "Yes",
}
es_data['did_create_portfolio'] = es_data['did_create_portfolio'].replace(did_create_portfolio_dict)

# Adding "other" job hunt blocker to a new column.

job_hunt_blockers_list = [
  "Nunca quise encontrar un trabajo como desarrollador",
  "Aprendí que el desarrollo web no es para mí",
  "Falta de motivación",
  "Frustración con el proceso",
  "Sentirse abrumado (demasiadas cosas que hacer",
  # the next two relate to the previous entry with a ","
  "demasiadas cosas que aprender",
  "etc.)",

  "No vi muchas ofertas de trabajo",
  "Falta de confianza en que encontraría un trabajo",
  "Procrastinación (quiero",
  # the next one relate to the previous entry with a ","
  "pero sigo retrasándolo y posponiéndolo)",

  "Vida personal (Me estoy centrando en otros aspectos de mi vida personal)",
  "Falta de orientación / no sé por dónde empezar",
]

# Function to find the unknown item
def find_unknown(blocker_str, valid_list):
    if pd.isna(blocker_str):
      return None
    items = [item.strip() for item in blocker_str.split(',')]
    unknowns = [item for item in items if item not in valid_list]
    #todo later, translate other
    return unknowns[0] if unknowns else None

# Apply the function to each row
es_data['other_job_hunt_blocker'] = es_data['job_hunt_blockers'].apply(lambda x: find_unknown(x, job_hunt_blockers_list))



# job_hunt_blockers replace

# NOTE: removing commas and other characters and simplifying response for simpler EDA later
job_hunt_blockers_dict = {
  "Nunca quise encontrar un trabajo como desarrollador": "I never wanted to find a job as a developer",
  "Aprendí que el desarrollo web no es para mí": "I learned that web dev is not for me",
  "Falta de motivación": "Lack of motivation",
  "Frustración con el proceso": "Frustration with the process",
  "Sentirse abrumado (demasiadas cosas que hacer": "Feeling overwhelmed",
  # the next two relate to the previous entry with a ","
  "demasiadas cosas que aprender": "",
  "etc.)": "",

  "No vi muchas ofertas de trabajo": "Did not see many job offers",
  "Falta de confianza en que encontraría un trabajo": "Lack of confidence",
  "Procrastinación (quiero": "Procrastination",
  # the next one relate to the previous entry with a ","
  "pero sigo retrasándolo y posponiéndolo)": "",

  "Vida personal (Me estoy centrando en otros aspectos de mi vida personal)": "Focusing on Personal life",
  "Falta de orientación / no sé por dónde empezar": "Lack of guidance",
}

# the lambda will split the substrings and apply the replace for each one
def filter_empty(elem):
  return elem != ""

es_data['job_hunt_blockers'] = es_data['job_hunt_blockers'].apply(
    lambda string: ', '.join(filter(filter_empty, [job_hunt_blockers_dict.get(substring.strip(), "Other") for substring in string.split(',')]))
    if pd.notnull(string) else string
)
# NOTE: adding "Other" instead of substring.strip() as a second argument of .get to group other responses. Later can be a different column.

# Adding "other" job success factor to a new column.

job_success_factor_list = [
  "Buena comunicación y habilidades blandas",
  "Buenas habilidades de código",
  "Un portafolio robusto o proyectos personales",
  "Networking y contactos",
  "Practiqué mucho para entrevistas técnicas (por ejemplo",
  # the next one relate to the previous entry with a ","
  "desafíos de codificación)",

  "Practiqué mucho para entrevistas no técnicas (por ejemplo",
  # the next three relate to the previous entry with a ","
  "preguntas de comportamiento",
  "explicación de proyectos",
  "etc.)",

  "Disciplina y trabajo duro",
  "Persistencia (me rechazaron muchas veces",
  # the next one relate to the previous entry with a ","
  "pero seguí postulando hasta que lo logré)",
]

# Function to find the unknown item
def find_unknown(factor_str, valid_list):
    if pd.isna(factor_str):
      return None
    items = [item.strip() for item in factor_str.split(',')]
    unknowns = [item for item in items if item not in valid_list]
    #todo later, translate other
    return unknowns[0] if unknowns else None

# Apply the function to each row
es_data['other_job_success_factor'] = es_data['job_success_factors'].apply(lambda x: find_unknown(x, job_success_factor_list))

# job_success_factors replace

# NOTE: removing commas and other characters and simplifying response for simpler EDA later. Repeating those not to be changed so the "other" is removed when not found.
job_success_factors_dict = {
  "Buena comunicación y habilidades blandas": "Good communication and soft skills",
  "Buenas habilidades de código": "Good coding skills",
  "Un portafolio robusto o proyectos personales": "Strong portfolio or personal projects",
  "Networking y contactos": "Networking and contacts",
  "Practiqué mucho para entrevistas técnicas (por ejemplo": "I practiced a lot for technical interviews",
  # the next one relate to the previous entry with a ","
  "desafíos de codificación)": "",

  "Practiqué mucho para entrevistas no técnicas (por ejemplo": "I practiced a lot for non-technical interviews",
  # the next three relate to the previous entry with a ","
  "preguntas de comportamiento": "",
  "explicación de proyectos": "",
  "etc.)": "",

  "Disciplina y trabajo duro": "Discipline and hard work",
  "Persistencia (me rechazaron muchas veces": "Persistance",
  # the next one relate to the previous entry with a ","
  "pero seguí postulando hasta que lo logré)": "",
}

# the lambda will split the substrings and apply the replace for each one

def filter_empty(elem):
  return elem != ""

es_data['job_success_factors'] = es_data['job_success_factors'].apply(
    lambda string: ', '.join(filter(filter_empty, [job_success_factors_dict.get(substring.strip(), "Other") for substring in string.split(',')]))
    if pd.notnull(string) else string
)

# NOTE: adding "Other" instead of substring.strip() as a second argument of .get to group other responses. Later can be a different column.

es_data.to_excel("../private_data/data/survey_results_es/es_data_translated.xlsx", index=False)
