==> notebooks/04_eda_general.ipynb <==
import sys
import os
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import Markdown

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "../scripts/eda/")))
from group_by_column_vs_has_dev_job import group_by_column_vs_has_dev_job

data = pd.read_excel("../private_data/data/cleaned_data/cleaned_data.xlsx")

## Filter students that found a job but it's not related to the bootcamp

For now, we're considering only students that found a dev job AND students that did not find a job. <br>

ie. for the analysis, we're not considering students that found a job but it's not related to the bootcamp (at least for now, since we don't have relevant data about them)



#
# IMPORTANT: for this analysis we're using only students that found a dev job AND students that didn't find a job.
#
# (ie. we're not considering students that found a job but it's not related to the bootcamp)
#

data = data[(data["post_bootcamp_situation"] != "I found a job, but is not related to the bootcamp")]



## EDA

#
# students that create projects, have higher chances of getting a job. new_projects_qty
#

# all students
display(Markdown("<br>"))
display(Markdown("### Number of projects created"))
students_2023_and_2024 = data[data["cohort_start_month"].str.startswith(("2023", "2024"))]
group_by_column_vs_has_dev_job(data, "new_projects_qty")



#
# students that create a portfolio, have higher chances of getting a job
#
display(Markdown("<br>"))
display(Markdown("### Portfolio"))
students_2023_and_2024 = data[data["cohort_start_month"].str.startswith(("2023", "2024"))]
group_by_column_vs_has_dev_job(data, "did_create_portfolio")

#
# amount of interviews to get a dev job
#

# students_2022
display(Markdown("<br>"))
display(Markdown("### students_2022:"))
students_2022 = data[data["cohort_start_month"].str.startswith(("2022"))]
r1 = group_by_column_vs_has_dev_job(students_2022, "company_interview_qty")
display(r1)

# students_2023
display(Markdown("<br>"))
display(Markdown("### students_2023:"))
students_2023 = data[data["cohort_start_month"].str.startswith(("2023"))]
r2 = group_by_column_vs_has_dev_job(students_2023, "company_interview_qty")
display(r2)

# students_2024
display(Markdown("<br>"))
display(Markdown("### students_2024:"))
students_2024 = data[data["cohort_start_month"].str.startswith(("2024"))]
r3 = group_by_column_vs_has_dev_job(students_2024, "company_interview_qty")
display(r3)


#
# students with previous knowledge have a higher chance of getting a job
#
display(Markdown("<br>"))
display(Markdown("### Previous coding knowledge"))
group_by_column_vs_has_dev_job(data, "pre_bootcamp_coding_experience")
