==> notebooks/04_eda_job_hunt_blockers.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

data = pd.read_excel("../private_data/data/cleaned_data/cleaned_data.xlsx")

data.columns

# only students with no dev job and not actively looking for a job anymore (does not count 7 students that took right path by mistake)
students_with_blockers = data[data["job_hunt_blockers"].notnull()] 

students_with_blockers["ct_student_id"].count()

# Step-by-step breakdown
all_blockers = students_with_blockers['job_hunt_blockers'] \
    .str.split(',') \
    .explode() \
    .str.strip()

# Count frequency
all_blockers.value_counts()
