==> notebooks/04_eda_job_success_factors.ipynb <==
import pandas as pd
# import matplotlib.pyplot as plt
# import seaborn as sns

data = pd.read_excel("../private_data/data/cleaned_data/cleaned_data.xlsx")

# only students with dev job or those 7 that went on the right path by mistake
students_with_success_factors = data[data["job_success_factors"].notnull()] 

students_with_success_factors["ct_student_id"].count()

# Step-by-step breakdown
all_success_factors = students_with_success_factors['job_success_factors'] \
    .str.split(',') \
    .explode() \
    .str.strip()

# Count frequency
all_success_factors.value_counts()

for val in students_with_success_factors[students_with_success_factors["other_job_success_factor"].notna()]["other_job_success_factor"]:
    print(val)

test = students_with_success_factors[students_with_success_factors["other_job_success_factor"].isna()]
test2 = students_with_success_factors[students_with_success_factors["other_job_success_factor"].notna()]

print(test.shape)
print(test2.shape)
